change server & add proxy_local & proxy_server
