+++ docs/design.md
@@ -20,8 +20,8 @@
 
 ### 设计方案
 #### PAC文件
-- 通过javascript实现的代理配置文件
-- 文件中可以设置哪些域名需要经过代理
+- 构建自定义的pac文件
+- 通过pac文件实现浏览器或操作系统的请求转发
 
 <br><br>
   
@@ -30,8 +30,8 @@
 2. 请求通过代理域名判断，需要走代理的走代理服务，不需要走代理的直接发起请求
 3. 请求以Socks5协议转发到代理服务客户端
 4. 本地代理服务将请求的域名和数据读取出来，将数据部分加密通过与服务端之间的连接转发出去
-5. 本地代理服务等待代理服务端的应答数据
-6. 客户端解密应答数据
+5. 本地代理服务等待代理服务服务端的应答数据
+6. 客户端收到应答数据并解密应答数据
 7. 应答数据被转发给客户端
 
 <br><br>
@@ -44,25 +44,31 @@
 - 启动服务
 - 循环等待请求
 - 接收请求
- - 连接被包装为客户实例
- - 创建读取协程
- - 创建写入协程
-- 读取的请求写入被加密，通过channel转发给代理服务
+ - 创建读取循环协程
+ - 创建写入循环协程
+- 读取请求的数据部分和请求头部分，并将请求数据加密转发给服务端
+- 读取协程等待服务端返回应答数据，将数据解密后通过channel写入客户端连接socket
   <br /><br />
 
 
 ### 模块化
   - 支持多种协议模块
+    - HTTPS
+    - Socks5 （优先）
   - 支持多种加密算法
+    - AES-256
   - 多种环境配置
+    - 客户端
+    - 服务端
   - 日志模块
+    - 独立日志模块
   <br /><br />
 
 
 ### 客户端和服务端连接
 #### 两种实现
-- 连接池
-- 每次新建连接
+- 连接池    （升级迭代）
+- 每次新建连接 （初期）
 
 #### 连接池的好处
 - 减少了中间连接环节的延迟
@@ -79,4 +85,17 @@
 
 ## 代理服务服务端
 #### 服务端代理服务内部逻辑 
-- 
+- 设置配置项
+- 初始化配置实例
+- 处理配置参数
+- 实例化服务
+- 启动服务
+- 循环等待请求
+- 接收请求
+ - 创建读取循环协程
+ - 创建写入循环协程
+- 读取请求的数据部分和请求头部分，解密请求体，通过请求体解析外部请求资源
+- 建立到外部的连接，通过连接转发请求体给对方，等待对方应答
+- 收到应答数据，将应答数据加密通过channel通知写入客户端请求socket
+
+
